[BUGFIX] Move url to Constants.kt
