+++ app/src/main/java/com/kazemieh/www/shop/ui/screens/home/ShowCaseSection.kt
@@ -17,6 +17,14 @@ import com.kazemieh.www.shop.ui.theme.LocalSpacing
 import com.kazemieh.www.shop.ui.theme.RoundedIconBox
 import com.kazemieh.www.shop.ui.theme.amber
 import com.kazemieh.www.shop.ui.theme.grayCategory
+import com.kazemieh.www.shop.util.Constants.AUCTION_URL
+import com.kazemieh.www.shop.util.Constants.DIGIJET_URL
+import com.kazemieh.www.shop.util.Constants.DIGIPAY_URL
+import com.kazemieh.www.shop.util.Constants.DIGIPLUS_URL
+import com.kazemieh.www.shop.util.Constants.GIFT_CARD_URL
+import com.kazemieh.www.shop.util.Constants.MORE_URL
+import com.kazemieh.www.shop.util.Constants.PINDO_URL
+import com.kazemieh.www.shop.util.Constants.SHOPPING_URL
 
 @Composable
 fun ShowCaseSection(navController: NavController) {
@@ -39,8 +47,7 @@ fun ShowCaseSection(navController: NavController) {
                 title = stringResource(id = R.string.my_shop_jet),
                 onClick = onBoxClick(
                     navController = navController,
-                    url = "https://www.kazemieh.com"
-//                    url = "https://www.digikalajet.com/user/address"
+                    url = DIGIJET_URL
                 )
             )
             RoundedIconBox(
@@ -48,7 +55,7 @@ fun ShowCaseSection(navController: NavController) {
                 title = stringResource(id = R.string.digi_style),
                 onClick = onBoxClick(
                     navController = navController,
-                    url = "https://www.digistyle.com/sale-landing/?utm_source=digikala&utm_medium=circle_badge&utm_campaign=style&promo_name=style&promo_position=circle_badge"
+                    url = AUCTION_URL
                 )
             )
             RoundedIconBox(
@@ -56,7 +63,7 @@ fun ShowCaseSection(navController: NavController) {
                 title = stringResource(id = R.string.digi_pay),
                 onClick = onBoxClick(
                     navController,
-                    url = "https://www.digikala.com/my-digipay/?promo_name=my-digipay&promo_position=circle_badge"
+                    url = DIGIPAY_URL
                 ),
             )
 
@@ -66,7 +73,7 @@ fun ShowCaseSection(navController: NavController) {
                 bgColor = MaterialTheme.colorScheme.amber,
                 onClick = onBoxClick(
                     navController,
-                    url = "https://www.pindo.ir/?utm_source=digikala&utm_medium=circle_badge&utm_campaign=pindo&promo_name=pindo&promo_position=circle_badge"
+                    url = PINDO_URL
                 ),
             )
 
@@ -84,28 +91,36 @@ fun ShowCaseSection(navController: NavController) {
                 title = stringResource(id = R.string.digi_shopping),
                 onClick = onBoxClick(
                     navController,
-                    url = "https://www.digikala.com/landing/gift-card-landing/?promo_name=gift_landing&promo_position=circle_badge"
+                    url =SHOPPING_URL
                 ),
 
                 )
             RoundedIconBox(
                 image = painterResource(id = R.drawable.giftcard),
                 title = stringResource(id = R.string.gift_card),
-                onClick = {}
+                onClick =onBoxClick(
+                    navController,
+                    url = GIFT_CARD_URL
+                )
+
             )
             RoundedIconBox(
                 image = painterResource(id = R.drawable.digiplus),
                 title = stringResource(id = R.string.digi_plus),
                 onClick = onBoxClick(
                     navController,
-                    url = "https://www.digikala.com/plus/landing/?promo_name=plus&promo_position=circle_badge"
+                    url = DIGIPLUS_URL
                 ),
             )
             RoundedIconBox(
                 image = painterResource(id = R.drawable.more),
                 title = stringResource(id = R.string.more),
                 bgColor = MaterialTheme.colorScheme.grayCategory,
-                onClick = {}
+                onClick =  onBoxClick(
+                    navController,
+                    url = MORE_URL
+                )
+
             )
 
         }
