--- app/src/main/java/com/kazemieh/www/shop/util/Constans.kt
@@ -1,17 +1,0 @@
-package com.kazemieh.www.shop.util
-
-import com.kazemieh.www.shop.BuildConfig
-
-object Constants {
-
-    const val ENGLISH_LANG = "en"
-    const val PERSIAN_LANG = "fa"
-
-    const val DATASTORE_NAME = "MAY_SHOP_DATA_STORE"
-    var USER_LANGUAGE = "USER_LANGUAGE"
-
-    const val NETWORK_TIME_OUT_SECOND = 60L
-    const val BASE_URL = "https://dig-za0p.onrender.com"
-    const val API_KEY = BuildConfig.X_API_KEY
-
-}
